fixed : 'CustomHTML2Text' is not defined
